Refactor: 문서 링크에 필요한 유틸리티 함수 분리
